Refactor ListCells and ListCellFactories into views package
